Extract Environment/EnvironmentOperation into container_schema.rs
Step 1 of splitting containers.rs into focused container_*.rs modules
(established naming convention: container_inspection, _lifecycle,
_logs, _routes, _runtime, _validation already exist and containers.rs
already re-exports from them). Moves the Environment struct (68
fields), EnvironmentOperation, and their ~45 default_*() helper
functions. Re-exported as `pub use crate::container_schema::
{Environment, EnvironmentOperation};` so none of the 13+ files that
import `containers::Environment` need any changes. 3446 -> 3157
lines.
